requirements.txt
python-telegram-bot>=21.0
boto3>=1.34
pyyaml>=6.0
httpx>=0.27
pydub>=0.25
filelock>=3.13
sarvamai>=0.1.0
